# api/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.IncidentDTO:
    properties:
      created_at:
        type: string
      days_ago:
        type: integer
      distance_km:
        type: number
      latitude:
        type: number
      longitude:
        type: number
      report_id:
        type: string
      risk_topic:
        type: string
      risk_type:
        type: string
      weight_factor:
        type: number
    type: object
...
  dto.NavigateToSavedLocationRequest:
    properties:
      current_lat:
        type: number
      current_lon:
        type: number
    required:
    - current_lat
    - current_lon
    type: object
...
  dto.RiskTopicResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      icon_url:
        type: string
      id:
        type: string
      name:
        type: string
      risk_type_id:
        type: string
      updated_at:
        type: string
    type: object
  dto.RiskTopicsListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.RiskTopicResponse'
        type: array
    type: object
  dto.RiskTypeResponse:
    properties:
      created_at:
        type: string
      default_radius:
        type: integer
      description:
        type: string
      icon_url:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
  dto.RiskTypesListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.RiskTypeResponse'
        type: array
    type: object
  dto.SafeRouteResponse:
    properties:
      calculated_at:
        type: string
      destination_lat:
        type: number
      destination_lon:
        type: number
      distance_km:
        type: number
      estimated_duration_minutes:
        type: integer
      id:
        type: string
      incident_count:
        type: integer
      incidents:
        items:
          $ref: '#/definitions/dto.IncidentDTO'
        type: array
      origin_lat:
        type: number
      origin_lon:
        type: number
      risk_level:
        type: string
      safety_score:
        type: number
      waypoints:
        items:
          $ref: '#/definitions/dto.WaypointDTO'
        type: array
    type: object
...
  dto.WaypointDTO:
    properties:
      latitude:
        type: number
      longitude:
        type: number
      sequence:
        type: integer
    type: object
...
  util.ErrorResponse:
    properties:
      error:
        properties:
          code:
            type: integer
          message: {}
        type: object
      success:
        type: boolean
    type: object
info:
  contact:
    email: support@riskplace.ao
    name: API Support
    url: http://www.riskplace.ao
  description: |-
    This is the API documentation for the Risk Place Angola application.

    ## Environments
    - **Development**: https://risk-place-angola-904a.onrender.com
    - **Local**: http://localhost:8000

    ## WebSocket Integration Guide
    For detailed instructions on integrating WebSocket in mobile applications, please refer to our [WebSocket Notification Guide](https://github.com/risk-place-angola/backend-risk-place/blob/develop/docs/MOBILE_WEBSOCKET_INTEGRATION.md)
  title: Risk Place Angola API
  version: 1.0.0
paths:
...
  /risks/topics:
    get:
      consumes:
      - application/json
      description: Retrieve risk topics, optionally filtered by risk_type_id query
        parameter.
      parameters:
      - description: Filter by risk type ID (UUID)
        in: query
        name: risk_type_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.RiskTopicsListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ErrorResponse'
      summary: List risk topics.
      tags:
      - risks
...
  /risks/types:
    get:
      consumes:
      - application/json
      description: Retrieve all available risk types with their default radius.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.RiskTypesListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/util.ErrorResponse'
      summary: List all risk types.
      tags:
      - risks
...
  /routes/navigate-home:
    post:
      consumes:
      - application/json
      description: Calculate a safe route from current location to the user's saved
        home address
      parameters:
      - description: Current location coordinates
        in: body
        name: location
        required: true
        schema:
          $ref: '#/definitions/dto.NavigateToSavedLocationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Safe route calculated successfully
          schema:
            $ref: '#/definitions/dto.SafeRouteResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "401":
          description: Unauthorized - missing or invalid JWT token
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "404":
          description: Home address not configured
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "500":
          description: Failed to calculate route
          schema:
            $ref: '#/definitions/util.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Calculate safe route to home address
      tags:
      - routes
  /routes/navigate-work:
    post:
      consumes:
      - application/json
      description: Calculate a safe route from current location to the user's saved
        work address
      parameters:
      - description: Current location coordinates
        in: body
        name: location
        required: true
        schema:
          $ref: '#/definitions/dto.NavigateToSavedLocationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Safe route calculated successfully
          schema:
            $ref: '#/definitions/dto.SafeRouteResponse'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "401":
          description: Unauthorized - missing or invalid JWT token
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "404":
          description: Work address not configured
          schema:
            $ref: '#/definitions/util.ErrorResponse'
        "500":
          description: Failed to calculate route
          schema:
            $ref: '#/definitions/util.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Calculate safe route to work address
      tags:
      - routes
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
  DeviceID:
    description: Unique device identifier for tracking and analytics.
    in: header
    name: X-Device-ID
    type: apiKey
swagger: "2.0"